requirements.txt
ultralytics==8.2.0
opencv-python==4.9.0.80
numpy<2.0.0
PyYAML==6.0.1
matplotlib==3.8.4
jupyter==1.0.0
pytest==8.1.1
tqdm==4.66.2
